Technically I could have written the code in C
I had tried to create a version that would use the C++ unordered_map, which worked for small values, but bugged out weirdly for large values.
Thus I created a table from the top down with the information needed to solve it. This is faster than the bottom-up process shown in class because not all values would need to be discovered throughout the table (the cost of this process being that more memory would be needed to store all the values that could have been used; but this is where the unordered_map class was supposed to come into play because I could instead have the key for a value of ways to get to the same amount of change be calculated from a string containing the first number of the index, followed by a comma, followed by the second number. This table could then dynamically grow to fit all of the needed information.